# docs/swagger.yaml
basePath: /
definitions:
...
  httpapi.createSleepProfileRequest:
    properties:
      baby_id:
        type: string
      night_window:
        $ref: '#/definitions/httpapi.nightWindowRequest'
      timezone:
        type: string
    type: object
  httpapi.editSleepSessionRequest:
    properties:
      started_at:
        type: string
      stopped_at:
        type: string
    type: object
  httpapi.errorResponse:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
...
  httpapi.nightWindowRequest:
    properties:
      end_hour:
        type: integer
      end_minute:
        type: integer
      start_hour:
        type: integer
      start_minute:
        type: integer
    type: object
  httpapi.sleepSessionResponse:
    properties:
      baby_id:
        type: string
      classification:
        type: string
      duration_seconds:
        type: number
      id:
        type: string
      started_at:
        type: string
      stopped_at:
        type: string
    type: object
  httpapi.startSleepRequest:
    properties:
      started_at:
        type: string
    type: object
  httpapi.startSleepResponse:
    properties:
      id:
        type: string
      started_at:
        type: string
    type: object
  httpapi.stopSleepRequest:
    properties:
      stopped_at:
        type: string
    type: object
  httpapi.stopSleepResponse:
    properties:
      classification:
        type: string
      id:
        type: string
      started_at:
        type: string
      stopped_at:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Baby sleep tracking API for families.
  title: Keklik API
  version: "1.0"
paths:
...
  /sleep-profiles:
    post:
      consumes:
      - application/json
      parameters:
      - description: Sleep profile configuration
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/httpapi.createSleepProfileRequest'
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpapi.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpapi.errorResponse'
      security:
      - BearerAuth: []
      summary: Create sleep profile
      tags:
      - sleep
  /sleep-sessions:
    get:
      parameters:
      - description: 'History window: today, 7d (default), or 14d'
        in: query
        name: period
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/httpapi.sleepSessionResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpapi.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpapi.errorResponse'
      security:
      - BearerAuth: []
      summary: Get sleep history
      tags:
      - sleep
    post:
      consumes:
      - application/json
      parameters:
      - description: Optional explicit start time (defaults to now)
        in: body
        name: body
        schema:
          $ref: '#/definitions/httpapi.startSleepRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/httpapi.startSleepResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpapi.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpapi.errorResponse'
        "409":
          description: Active sleep session already exists
          schema:
            $ref: '#/definitions/httpapi.errorResponse'
      security:
      - BearerAuth: []
      summary: Start sleep session
      tags:
      - sleep
  /sleep-sessions/{id}:
    delete:
      parameters:
      - description: Sleep session UUID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpapi.errorResponse'
      security:
      - BearerAuth: []
      summary: Delete sleep session
      tags:
      - sleep
    patch:
      consumes:
      - application/json
      parameters:
      - description: Sleep session UUID
        in: path
        name: id
        required: true
        type: string
      - description: Fields to update (at least one required)
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/httpapi.editSleepSessionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httpapi.sleepSessionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpapi.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpapi.errorResponse'
      security:
      - BearerAuth: []
      summary: Edit sleep session
      tags:
      - sleep
  /sleep-sessions/active:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Optional explicit stop time (defaults to now)
        in: body
        name: body
        schema:
          $ref: '#/definitions/httpapi.stopSleepRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httpapi.stopSleepResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpapi.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpapi.errorResponse'
      security:
      - BearerAuth: []
      summary: Stop active sleep session
      tags:
      - sleep
securityDefinitions:
  BearerAuth:
    description: Session token obtained from the auth endpoints.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"